fix: #1086 make JWT jti consistent with what's in system
The previous code generates the jws before saving the token records by
then a new key ID is generated, hence causing the ID in the JWT
different with what's saved in Terrakube database.

This change addresses this by moving save operation before the JWS
generation and reuses the ID generated by JPA/Spring in the JWT, in an
exception, the record in the Terrakube database will be removed.