Add proxy support to GitHubTokenService (#931)
Added support for proxy in the GitHubTokenService. The code now checks if a proxy host is set in system properties and if so, uses it to create the HttpClient. If no proxy host is specified, it falls back to the default settings. Additionally, log messages were introduced to aid in debugging.