feat: Single GitHub App token per installation
It's possible and quite common that one single GitHub App is used across
different VCS providers in multiple Terrkube organisations, in which
case the access token can be shared in the reason that the app permissions
of each installation are configured against the GitHub app itself, which
has nothing to do with Terrakube organisation structure.

Such sharing also reduces the tokens kept in the database and scheduled
jobs to refresh these tokens.